libdrgn: python: cast enums when wrapping with Python call
The "k" format expects an unsigned long, so make sure we cast C enums to
the proper type. This probably doesn't matter for x86 in practice, but
it's better to be safe.

+++ libdrgn/python/program.c
@@ -217,7 +217,8 @@ static struct drgn_error *py_type_find_fn(enum drgn_type_kind kind,
 	PyObject *type_obj;
 
 	gstate = PyGILState_Ensure();
-	kind_obj = PyObject_CallFunction(TypeKind_class, "k", kind);
+	kind_obj = PyObject_CallFunction(TypeKind_class, "k",
+					 (unsigned long)kind);
 	if (!kind_obj) {
 		err = drgn_error_from_python();
 		goto out_gstate;
